mv auto_run_name to common
